Split frontend logic from backend logic.
This is a very big refactoring which incorporated a lot of design changes
in order to accomplish the frontend vs. backend separation.

Stabilized a public API.

Fixed a bunch of bugs.

Implemented a whole bunch of unit tests and integration tests.

Wrote a completely new documentation.

A lot of stuff that I didn't mention ... See the remove_frontend_branch
log history for more details.