Check if given compiler is available on the system.
It might be that some clients will not have all dependencies installed
right away, e.g. clang compiler suite, in which case we will gracefully
fail instead of generating an exception.

It should possible to expose this configuration to clients to hard-code
it themselves if they choose not to rely on installing yet another dep.