Implement integration tests profiling.
Each of the tests which are part of the integration
test suite are now automatically profiled with the means
of cProfile, a Python profiler.

Each test profile stats are sorted by cummulative time of functions
that have been run during the test. When all tests have been run, they
will be sorted by their total execution time which is handy because we
can identify biggest offenders more easily.

Because cProfile is not able to track function calls across multiple-process
boundary, profiling results will not actually be as detailed as one would
expect. However, we still gain from it because we get figures on what happens
on a big-picture scale which we can use to identify when something goes wrong
in lower layers. More importantly, it enables us to test the implementation on
the system-level a.k.a. what will really happen when we start using services
in a way a frontend (IDE) would use it. This discovers bugs which are hard or
impossible to catch with standard unit-tests.