Keeping cache up-to-date in code-completion use-case probably doesn't make too much sense.
ClangParser.parse() is an overkill here because it will _reparse_ the TUnit once it
identifies that the TUnit it got from cache is outdated (by looking at the mtime). This is
an overhead in code-completion use-case because libclang will reparse the TUnit anyway so we
would actually end up wasting resources.