Implement support to handle arbitrary build-targets.
Up to this patch, it was the frontend responsibility to detect correct
path to compilation-db/compile-flags configuration and to forward
it to the Cxxd services which required this kind of information.

Due to the fact that Cxxd recently implemented support for
server-side configuration through .cxxd_config.json means,
it is natural to move this responsibility to the server which
enables users to arbitrarily populate a 'configuration' section
in .cxxd_config.json with relevant information about build-targets
for the project they are running Cxxd server on.

This approach is much more flexible because it allows users to tweak
configuration settings without hard-coding them into the frontend code,
like it was the case previously. cxxd-vim for example had search-paths
(for discovery of compilation-db/compile-flags) hard-coded in its plugin
Vimscript code. That is not the most portable way of doing it nor it is
possible to have 1 universal search-path solution for every project out
there.

Also, having a possibility to define multiple build-targets, which
previously was not the case, we can feed the server with the appropriate
configuration. E.g. if we want server to "understand" debug-build code,
then we would have to feed it with all the debug-relevant build-flags,
include directories etc. (whatever we use when building for debug).
This is now simple as configuring the server with respective debug-build
configuration. Shall we want to switch to release-build, we shall feed
the server with another release-build configuration. It's that trivial.
Current limitation though is that server needs to be restarted in order for
that change to take the effect but it will be considered to make it possible
to make a switch in runtime (shouldn't be too hard but requires possibility
to restart services which we currently don't have).

Configuration of build-target settings can be done:
  1. Auto-magically or
  2. (recommended) Through explicit configuration.

Auto-magically:
  * This mode will be triggered in:
      1. Absence of .cxxd_config.json file.
      2. Absence of 'configuration' section of .cxxd_config.json file.
      3. Absence of 'type' information within 'configuration' section.
  * Success of this mode depends on the *hard-coded* paths within the
    Cxxd-server code.
  * It is NOT recommended to use this mode. It is here for convenience
    and for most simplest cases (e.g. when you're trying out things
    but don't necessarily have a full-blown project configuration).

Explicit:
  * This mode requires presence of:
      1. .cxxd_config.json file in the root of the project directory.
      2. Valid build-target configuration.
  * Valid configuration requires:
      1. 'configuration' section
      2. Valid 'type' within 'configuration' section. It can be one of
         the following:
           'compilation-database' | 'compile-flags' | 'auto-discovery'
      3. Build-targets configuration.
	   * If 'type' is set to 'compilation-database' OR 'compile-flags'
             then user shall provide:
               * a new field named 'compilation-database' or 'compile-flags'
               * 'target' field within previous field which shall contain
                 as many arbitrary pairs of
                     '<build-target-name>' : '<build-target-config-path>'
                 as project may want to use. Path shall point to the directory
                 where configuration for relevant build-target is located.
           * If 'type' is set to 'auto-discovery' then user must provide:
               * 'search-paths' field
                   * This field shall contain directory paths which server
                     shall use in order to discover configuration file.
           * Otherwise, configuration is invalid and server will fallback
             to auto-magic discovery mode (not recommended).

Example of how build-target configuration may look like:
{
    "configuration" : {
        "type" : "compilation-database",
        "compilation-database" : {
            "target" : {
                "debug" : "../debug_build",
                "release" : "../release_build",
                "relwithdbginfo" : "../relwithdbginfo_build"
            }
        }
    }
}

Server then could be invoked with any of the targets defined in the
configuration file. If no compilation-database or no compile-flags
configuration file could be found, server will bail-out because of
the very limited parsing capabilities it would have in that case.