Add possibility of setting the libclang.so path through cxxd config.
This is both convenience and necessity for two reasons.

First one being that from time to time bugs do lurk into the libclang
stack or sometimes even some documented behavior is suddenly changed.
Recent example that I had to fix is an API break in clang_getCompileCommands,
in particular it changed the way how it handles the header files.
Another recent one is a bug in python libclang bindings which changed the
logic of loading the libclang library by assuming that the library will
be available _only_ under the name libclang-14.so. I think this is pretty
awkward. Both of this type of issues are hard to debug and they always
come as a surprise because merely update your system libraries from time
to time will get you into this type of issues.

There has been numerous bugs similar to the ones I described above and
which I already had to fix in the past so the second point being, perhaps
cxxd should start prepackaging the version of libclang.so and python
bindings which are known to cause no unsurprising effects. This will
result in more stable experience with the trade-off of having to bump
to the newer versions of libclang/python-bindings manually. In this case,
integration test suite will come handy.