Implement support for defining custom per-target build commands.
Now, we may define something like the following in our per-project
config file:

  ...
    "project-builder" : {
        "target": {
            "debug" : {
                "cmd": "ninja"
            },
            "relwithdbg": {
                "cmd": "ninja"
            },
            "relwithdbg-distcc-pump": {
                "cmd": "pump ninja -j $(distcc -j)"
            }
        }
    }
  ...

By using this feature, we will be able to follow up with another handy
feature and which is being able to run the build with appropriate
build-command which is deduced automagically depending on which target
is currently selected.

For example:
 (1) cxxd-start . relwithdbg
 (2) cxxd-build-auto

cxxd-build-auto will now be able to use the build-cmd that is parsed and
extracted from .cxxd_config.json file, in this case "ninja".

This will further enable frontends to define keyboard shortcuts or
similar actions and map them simply to this feature.