Add API for common and project-specific (if any) cpp extensions.
This will make it easier for services to decline the action/operation
in case the extension is not recognized. This will lower the
burden on the frontend side which now has to do these things manually
for each and every service.

Future work will make use of this API to implement the filtering.