Fix the clang_getCompileCommands API break.
Previously, when clang_getCompileCommands was invoked with a header file
it couldn't return any compile commands because header file is not a
translation unit so there was no related entry in the compilation database.
This was expected and which is why I implemented some sort of heuristics
which were providing the best compile command guesstimate for any header
file.

New and undocumented behavior of clang_getCompileCommands is now that
it will even for header files return some sort of a compile command which
it did not previously. The command that it will return, funnily, won't
work with the clang parser and parsing the header will result with an
"invalid arguments" error.

This patch leaves the old heuristics and makes a special-case for
header files in order to workaround the above mentioned issues.