Disallow indexing the modified source code file at the API level.
Previously, I had exposed this possibility through a public facing API
and also had the logic preventing the very same in the indexing
implementation. This doesn't make sense so let's get rid of the logic
and prevent doing it at the API level. I think this was a very early
artifact of initial public API design which went unnoticed for some time.