Extension for clang_getCompileCommands API break fix.
It appears that clang_getCompileCommands will return a non-empty compilation
command not only for header files but generally for all files that are
not part of the compilation database. Consequently this means that
actual translation units which are not yet part of the compilation
database will also get the special handling from clang_getCompileCommands.

This situation can easily happen when a new .cxx file is created but
it is not yet added as an entry to CMakeLists.txt. Because of that,
compile_commands.json will not contain it as entry.

It also can happen even if the file has been added to CMakeLists.txt but
the CMake build configuration has not been regenerated.

This fix addresses these issues.