Implement a disassembler service.
It is inspired by the idea of Godbolt but with the difference of having
a capability to be used on the actual projects (e.g. on a whole bunch of
different source-code files) and not only on the excerpts of your code
that you have copy-pasted into the Godbolt editor.

This means that this disassembler service does not operate on chunks of
an exemplary code but it operates on an actual binaries your code is
going to produce. What it will do is that it will try to match the symbol
you expressed the interest in on a source-code level, e.g. some function,
against the actual symbols that are present in the resulting binary. It
will do so by disassembling the binary and jumping to the symbol of
interest in the resulting disassembly.

Sometimes a symbol of interest won't even be present in the resulting
binary because it was either optimized out or inlined or sometimes even
merged with some other symbol. This normally happens because of the
compiler and/or linker optimizations and it is inevitable. What can also
happen is that the symbol of interest is not an actual symbol but rather
a meta-symbol such as function template or a class template, and therefore
cannot be present in such a form in a resulting binary.

In both of these cases, disassembly service will instead offer a list
which contains all of the similar symbols that we can jump to. In case
of function/class templates this approach works quite well. But in case
when the symbol is completely missing, we will need to fallback onto
manual heuristics of finding the interesting part of the disassembly.

Not having all the code isolated into a single translation-unit, like we
do in a Godbolt, this service gives us an ability to examine the codegen
on a project level rather than on a single source-code file level. This
is a much tougher challenge for your compiler and linker to solve, and
hence it may/will impact the resulting codegen. Service also provides
flexibility to pick any target (binary) available to your project you
might be interested in (e.g. unit-tests, google-benchmarks, etc.).

Whole service is implemented natively through nm + objdump + libclang
combination, and has no dependencies or whatsoever on Godbolt. It runs
completely locally within the cxxd server.

It also provides the convenience not having to c/p the excerpts into a
Godbolt editor, which may or may not be a sensitive content, or cutting
loose the project-specific dependencies so that you can successfully
compile the excerpt. This is sometimes time-consuming, especially in
large codebases.

To understand the disassembly better, service also provides a documentation
of x86-64 assembly instructions and which can be retrieved by the client
application (e.g. on hovering the mouse over some assembly instruction).
This part is implemented with the help of the script which Godbolt uses
to scrape the documentation from https://www.felixcloutier.com/x86/index.html.
I have tweaked the script to fit the purposes of cxxd better.

There're some disassembly-related knobs that can be tweaked through the
.cxxd_config.json configuration file. These include:
  (1) "targets-filter" - to filter out the uninteresting targets from
      your build directory. E.g. "targets-filter" : [".so", ".a", "CMake"]
  (2) "intermix-with-src-code" - to mix the disassembly with the source
      code. By default it is turned off.
  (3) "syntax" - to select the disassembly syntax. By default it is set
      to intel syntax but can be changed to AT&T if someone wants to.
  (4) "visualize-jumps" - to visualize the branches in disassembled code.
      By default this setting is turned on.
  (5) "binary" - Both nm and objdump custom binaries can be provided
      should there be a reason for it.

E.g.
 "disassembly" : {
        "objdump" : {
            "binary" : "",
            "intermix-with-src-code" : false,
            "visualize-jumps" : true,
            "syntax" : "intel"
        },
        "nm" : {
            "binary" : ""
        },
        "targets-filter" : [".so", ".a", "CMake"]
    }